Moved ShortCodeUniquenessHelper to ShortUrl\Helper namespace
