+++ module/Core/src/ShortUrl/Helper/ShortCodeUniquenessHelper.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Core\ShortUrl;
+namespace Shlinkio\Shlink\Core\ShortUrl\Helper;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Shlinkio\Shlink\Core\Entity\ShortUrl;
